Add default display for related content
Add a default display behaviour for the related content part. Display the
related items in a list shape using their summary display type.

Add a new property to the RelatedContentPart for a CollectionDisplayShape
that lets you override List with a shape of your own when you are
displaying the items.

Update Placement.info to hide the new display shape unless using a Detail
DisplayType.